feat(python): add responder and split modules into files
